requirements.txt
matplotlib
numpy
open3d
Pillow
pyrender
rerun
setuptools
tqdm
trimesh
huggingface-hub
wandb 
einops

# for app.py
onnxruntime
gradio
rembg
git+https://github.com/microsoft/MoGe.git